initial kdr implementation

+++ src/Shared.h
@@ -122,12 +122,12 @@ struct SquadStats {
     uint64_t totalDamageVsPlayers = 0;
     uint64_t totalStrikeDamageVsPlayers = 0;
     uint64_t totalCondiDamageVsPlayers = 0;
-    int totalKillsVsPlayers = 0;
-    double getKillDeathRatio() const {
+    uint32_t totalKillsVsPlayers = 0;
+    float getKillDeathRatio() const {
         if (totalDeathsFromKillingBlows == 0) {
-            return static_cast<double>(totalKills);
+            return static_cast<float>(totalKills);
         }
-        return static_cast<double>(totalKills) / totalDeathsFromKillingBlows;
+        return static_cast<float>(totalKills) / totalDeathsFromKillingBlows;
     }
     std::unordered_map<std::string, SpecStats> eliteSpecStats;
 };
@@ -144,13 +144,13 @@ struct TeamStats {
     uint64_t totalDamageVsPlayers = 0;
     uint64_t totalStrikeDamageVsPlayers = 0;
     uint64_t totalCondiDamageVsPlayers = 0;
-    int totalKillsVsPlayers = 0;
+    uint32_t totalKillsVsPlayers = 0;
     bool isPOVTeam = false;
-    double getKillDeathRatio() const {
+    float getKillDeathRatio() const {
         if (totalDeathsFromKillingBlows == 0) {
-            return static_cast<double>(totalKills);
+            return static_cast<float>(totalKills);
         }
-        return static_cast<double>(totalKills) / totalDeathsFromKillingBlows;
+        return static_cast<float>(totalKills) / totalDeathsFromKillingBlows;
     }
     std::unordered_map<std::string, SpecStats> eliteSpecStats;
     SquadStats squadStats;
